fix: wrap floor(random()*N) in f32() to prevent WGSL implicit i32→f32 conversion warnings across 24 variation files
floor() returns i32 in WGSL, which then gets compared with f32 literals or used in f32 arithmetic, producing browser console warnings. The fix pattern: floor(random() * N) → f32(floor(random() * N)).

Affected files: symBandG6/G7, symNetG4-G17, postPointSymmetryWf, nPolarVar, chaosCubesVar (4 occurrences), sattractorVar, juliaQVar, woggleVar, waffleVar (9 occurrences), bTransformVar, powBlockVar.

+++ packages/app/src/flame/variations/parametric/general/bTransformVar.tsx
@@ -69,7 +69,8 @@ export const bTransformVar = parametricVariation(
     let sigma =
       PI.$ - atan2(pos.y, pos.x + 1.0) - atan2(pos.y, 1.0 - pos.x) + P.rotate
     sigma =
-      sigma / P.power + ((2.0 * PI.$) / P.power) * floor(random() * P.power)
+      sigma / P.power +
+      ((2.0 * PI.$) / P.power) * f32(floor(random() * P.power))
 
     tau = select(tau - P.split, tau + P.split, pos.x >= 0.0)
