+++ packages/app/src/flame/variations/parametric/post/symNetG6.tsx
@@ -36,7 +36,7 @@ export const symNetG6 = parametricVariation(
     const hx = P.stepx * 0.5
     const hy = P.stepy * 0.5
     const sy2 = P.sepy * 0.5
-    const band = floor(random() * 4.0)
+    const band = f32(floor(random() * 4.0))
     if (band < 1.0) {
       return vec2f(pos.x - hx, pos.y - hy + sy2)
     }
